ADPS-1168 - fixed mr comments
