--- dev-support/resource-mapping-manager/conf/ranger-trino-security.xml
@@ -1,60 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<!--
-  Licensed to the Apache Software Foundation (ASF) under one or more
-  contributor license agreements.  See the NOTICE file distributed with
-  this work for additional information regarding copyright ownership.
-  The ASF licenses this file to You under the Apache License, Version 2.0
-  (the "License"); you may not use this file except in compliance with
-  the License.  You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
--->
-<?xml-stylesheet type="text/xsl" href="configuration.xsl"?>
-<configuration>
-    <property>
-        <name>ranger.plugin.trino.service.name</name>
-        <value>dev_trino</value>
-    </property>
-
-    <property>
-        <name>ranger.plugin.trino.policy.source.impl</name>
-        <value>org.apache.ranger.admin.client.RangerAdminRESTClient</value>
-    </property>
-
-    <property>
-        <name>ranger.plugin.trino.policy.rest.url</name>
-        <value>http://ranger:6080</value>
-    </property>
-
-    <!-- Chained plugin configurations -->
-    <property>
-        <name>ranger.plugin.trino.chained.services</name>
-        <value>dev_hive</value>
-    </property>
-
-    <property>
-        <name>ranger.plugin.trino.chained.services.dev_hive.impl</name>
-        <value>org.apache.ranger.hive.chained.trino.TrinoHiveChainedPlugin</value>
-    </property>
-
-    <property>
-        <name>ranger.plugin.hive.service.name</name>
-        <value>dev_hive</value>
-    </property>
-
-    <property>
-        <name>ranger.plugin.hive.policy.rest.url</name>
-        <value>http://ranger:6080</value>
-    </property>
-
-    <property>
-        <name>ranger.plugin.hive.policy.source.impl</name>
-        <value>org.apache.ranger.admin.client.RangerAdminRESTClient</value>
-    </property>
-</configuration>
