+++ distro/src/main/assembly/plugin-trino.xml
@@ -28,7 +28,7 @@
             <includes>
                 <include>org.apache.ranger:ranger-plugins-audit</include>
                 <include>org.apache.ranger:ranger-plugins-cred</include>
-                <include>org.apache.ranger:ranger-hive-plugin</include>
+                <include>org.apache.ranger:ranger-hive-chained-plugin-base</include>
                 <include>org.apache.ranger:ranger-plugins-common</include>
                 <include>org.apache.ranger:ranger-trino-plugin</include>
             </includes>
@@ -82,7 +82,7 @@
                     <include>org.apache.hadoop:hadoop-auth:jar:${hadoop.version}</include>
                     <include>org.apache.hadoop:hadoop-hdfs:jar:${hadoop.version}</include>
                     <include>org.apache.ranger:ranger-plugins-cred</include>
-                    <include>org.apache.ranger:ranger-hive-plugin</include>
+                    <include>org.apache.ranger:ranger-hive-chained-plugin-base</include>
                     <include>org.apache.ranger:credentialbuilder</include>
                     <include>org.codehaus.woodstox:stax2-api</include>
                     <include>com.fasterxml.woodstox:woodstox-core</include>
