+++ ranger-hive-chained-plugin-base/src/main/java/org/apache/ranger/hive/chained/plugin/HiveChainedPlugin.java
@@ -19,26 +19,9 @@
 
 package org.apache.ranger.hive.chained.plugin;
 
-import static org.apache.ranger.hive.chained.plugin.HiveChainedPluginConfigKeys.MAPPINGS_FILE_LOCATION_DEFAULT;
-import static org.apache.ranger.hive.chained.plugin.HiveChainedPluginConfigKeys.MAPPINGS_FILE_LOCATION_POSTFIX;
-import static org.apache.ranger.hive.chained.plugin.HiveChainedPluginConfigKeys.MAPPINGS_PERSIST_INTERVAL_DEFAULT;
-import static org.apache.ranger.hive.chained.plugin.HiveChainedPluginConfigKeys.MAPPINGS_PERSIST_INTERVAL_POSTFIX;
-import static org.apache.ranger.hive.chained.plugin.HiveChainedPluginConfigKeys.MAPPINGS_REFRESH_INTERVAL_DEFAULT;
-import static org.apache.ranger.hive.chained.plugin.HiveChainedPluginConfigKeys.MAPPINGS_REFRESH_INTERVAL_POSTFIX;
-import static org.apache.ranger.plugin.policyengine.RangerPolicyEngine.ANY_ACCESS;
-
-import java.io.IOException;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-import java.util.stream.Collectors;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.ranger.admin.client.RangerAdminClient;
-import org.apache.ranger.authorization.hive.authorizer.HiveAccessType;
 import org.apache.ranger.authorization.hive.authorizer.HiveObjectType;
-import org.apache.ranger.authorization.hive.authorizer.RangerHiveAccessRequest;
 import org.apache.ranger.authorization.hive.authorizer.RangerHiveResource;
 import org.apache.ranger.hive.chained.mapping.HiveEntity;
 import org.apache.ranger.hive.chained.mapping.HiveMappingFetcher;
@@ -46,14 +29,24 @@ import org.apache.ranger.hive.chained.mapping.HiveResourceMappingStore;
 import org.apache.ranger.hive.chained.mapping.TrieHiveResourceMappingStore;
 import org.apache.ranger.plugin.mapping.ResourceMappingFetcher;
 import org.apache.ranger.plugin.policyengine.RangerAccessRequest;
-import org.apache.ranger.plugin.policyengine.RangerAccessResource;
 import org.apache.ranger.plugin.policyengine.RangerPluginContext;
 import org.apache.ranger.plugin.policyengine.RangerResourceACLs;
 import org.apache.ranger.plugin.service.RangerBasePlugin;
-import org.apache.ranger.plugin.service.ResourceMappingChainedPlugin;
+
+import java.io.IOException;
+import java.util.Collections;
+import java.util.List;
+import java.util.Optional;
+
+import static org.apache.ranger.hive.chained.plugin.HiveChainedPluginConfigKeys.MAPPINGS_FILE_LOCATION_DEFAULT;
+import static org.apache.ranger.hive.chained.plugin.HiveChainedPluginConfigKeys.MAPPINGS_FILE_LOCATION_POSTFIX;
+import static org.apache.ranger.hive.chained.plugin.HiveChainedPluginConfigKeys.MAPPINGS_PERSIST_INTERVAL_DEFAULT;
+import static org.apache.ranger.hive.chained.plugin.HiveChainedPluginConfigKeys.MAPPINGS_PERSIST_INTERVAL_POSTFIX;
+import static org.apache.ranger.hive.chained.plugin.HiveChainedPluginConfigKeys.MAPPINGS_REFRESH_INTERVAL_DEFAULT;
+import static org.apache.ranger.hive.chained.plugin.HiveChainedPluginConfigKeys.MAPPINGS_REFRESH_INTERVAL_POSTFIX;
 
 @Slf4j
-public abstract class HiveChainedPlugin extends ResourceMappingChainedPlugin {
+public abstract class HiveChainedPlugin extends BaseHiveChainedPlugin {
     public static final String HIVE_SERVICE_TYPE = "hive";
 
     private final HiveResourceMappingStore resourceMappingsStore;
@@ -125,37 +118,11 @@ public abstract class HiveChainedPlugin extends ResourceMappingChainedPlugin {
         }
 
         return maybeHiveEntity
-            .map(entity -> toAccessRequests(entity, request))
+            .map(entity -> toHiveAccessRequests(toHiveResource(entity), request))
             .orElseGet(Collections::emptyList);
     }
-
-    protected List<HiveAccessType> getAccessMappings(String... mappings) {
-        return Arrays.stream(mappings)
-            .map(this::toAccessType)
-            .collect(Collectors.toList());
-    }
-
-    protected HiveAccessType toAccessType(String rawAccessType) {
-        return rawAccessType.equals(ANY_ACCESS)
-            // Hive plugin uses "USE" access type as synonym for "_any"
-            ? HiveAccessType.USE
-            : HiveAccessType.valueOf(rawAccessType.toUpperCase());
-    }
-
-    protected abstract Optional<AccessMappings> getAccessTypeMappings(HiveObjectType hiveObjectType);
 
     protected abstract Optional<String> getPathFromRequest(RangerAccessRequest request);
-
-    protected List<HiveAccessType> getHiveAccessTypes(HiveEntity entity, RangerAccessRequest request) {
-        List<HiveAccessType> accessTypes = getAccessTypeMappings(entity.getType())
-            .map(mappings -> mappings.getHiveAccessTypes(request))
-            .orElseGet(Collections::emptyList);
-
-        if (accessTypes.isEmpty()) {
-            log.warn("No access type mapping found for request: {}", request);
-        }
-        return accessTypes;
-    }
 
     protected HiveMappingFetcher newMappingFetcher(
         RangerAdminClient adminClient,
@@ -170,31 +137,6 @@ public abstract class HiveChainedPlugin extends ResourceMappingChainedPlugin {
             mappingsPersistInterval,
             targetService
         );
-    }
-
-    private List<RangerAccessRequest> toAccessRequests(HiveEntity entity, RangerAccessRequest srcRequest) {
-        return getHiveAccessTypes(entity, srcRequest)
-            .stream()
-            .map(accessType -> new RangerHiveAccessRequest(
-                toHiveResource(entity),
-                srcRequest.getUser(),
-                srcRequest.getUserGroups(),
-                srcRequest.getUserRoles(),
-                getActionDescription(srcRequest, accessType),
-                accessType,
-                null,
-                null
-            ))
-            .collect(Collectors.toList());
-    }
-
-    private String getActionDescription(RangerAccessRequest srcRequest, HiveAccessType hiveAccessType) {
-        return String.format("%s (%s) -> %s",
-            srcRequest.getAccessType(),
-            Optional.ofNullable(srcRequest.getResource())
-                .map(RangerAccessResource::getAsString)
-                .orElse(""),
-            hiveAccessType);
     }
 
     private RangerHiveResource toHiveResource(HiveEntity entity) {
@@ -215,18 +157,5 @@ public abstract class HiveChainedPlugin extends ResourceMappingChainedPlugin {
             nameSegments.get(1),
             nameSegments.get(2)
         );
-    }
-
-    protected static class AccessMappings {
-        private final Map<String, List<HiveAccessType>> accessTypeMappings;
-
-        public AccessMappings(Map<String, List<HiveAccessType>> accessTypeMappings) {
-            this.accessTypeMappings = accessTypeMappings;
-        }
-
-        public List<HiveAccessType> getHiveAccessTypes(RangerAccessRequest request) {
-            log.debug("Trying to get access type mappings for {}", request);
-            return accessTypeMappings.getOrDefault(request.getAccessType(), Collections.emptyList());
-        }
     }
 }
